修改dict

+++ dict-service-provider/src/main/java/com/moseeker/dict/service/impl/CityServices.java
@@ -254,7 +254,7 @@ public class CityServices {
         if(!StringUtils.isEmptyList(list)){
             List<Map<String,Object>> result=new ArrayList<>();
             for(Map<String,Object> data:list){
-                int level= (int) data.get("level");
+                byte level= (byte) data.get("level");
                 int code=(int) data.get("code");
                 if(level==1){
                     if(!codeList.contains(code)){
@@ -278,7 +278,7 @@ public class CityServices {
             List<Map<String,Object>> cityList=new ArrayList<>();
             for(Map<String,Object> data:list){
                 int coityCode= (int) data.get("code");
-                int level=(int)data.get("level");
+                byte level= (byte) data.get("level");
                 if(coityCode>code&&coityCode<code+10000&&level>1&&!codeList.contains(coityCode)){
                     cityList.add(data);
                 }
